backend/requirements.txt

fastapi
uvicorn
httpx
python-dotenv
youtube-transcript-api
